requirements.txt
ollama
fastapi
pydantic_settings
python-dotenv
azure-storage-blob